requirements.txt
flask
flask-cors
toml
urllib3
requests
colorama
fastlogging
Jinja2
mistletoe
markdownify
pdfminer.six
playwright
pytest-playwright
tiktoken
sqlmodel
keybert
GitPython
netlify-py
Markdown
xhtml2pdf
Flask-SocketIO
eventlet
duckduckgo-search
orjson
gevent
gevent-websocket
curl_cffi
PyYAML
tenacity
openai>=1.0.0
qdrant-client
sentence-transformers
python-dotenv
